+++ client/src/components/TabResize/DynamicTab.tsx
@@ -2,13 +2,13 @@
 import React, { ReactNode } from 'react';
 import Box from '@mui/material/Box';
 
-interface DynamicBoxProps {
+interface DynamicTabProps {
   children: ReactNode;
   sx?: object;
 }
 
-const DynamicBox: React.FC<DynamicBoxProps> = ({ children, sx }) => {
+const DynamicTab: React.FC<DynamicTabProps> = ({ children, sx }) => {
   return <Box sx={{ border: '1px solid black', padding: '8px', ...sx }}>{children}</Box>;
 };
 
-export default DynamicBox;
+export default DynamicTab;
