+++ client/src/components/TabResize/index.tsx
@@ -1,2 +1,2 @@
 // visualiser-debugger/components/TabResize/index.tsx
-export { default as DynamicBox } from './DynamicBoxes';
+export { default as DynamicBox } from './DynamicTabs';
